Fixes around Editor and Owner permissions checks
We are staying away from primitive roles as much as possible and since
recently we are no longer using the Owner role in our projects. That
resulted in InSpec errors because there were no longer any IAM
bindings with that role. With the following fix, InSpec will look for
any bindings related to editor/owner and, in case of any present, will
examine those.